+++ processor/src/test/fixtures/input/com/example/cascade_dispose/Suppressed1DisposeOnDeactivateCascadeDisposeFieldModel.java
@@ -1,0 +1,17 @@
+package com.example.cascade_dispose;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.CascadeDispose;
+
+@ArezComponent
+abstract class Suppressed1DisposeOnDeactivateCascadeDisposeFieldModel
+{
+  @SuppressWarnings( "Arez:ConflictingDisposeModel" )
+  @CascadeDispose
+  protected MyComponent _myElement;
+
+  @ArezComponent( allowEmpty = true, disposeOnDeactivate = true )
+  abstract static class MyComponent
+  {
+  }
+}